renderer: Move shader template to resources namespace.
