+++ libopenage/renderer/stages/world/CMakeLists.txt
@@ -2,5 +2,4 @@ add_sources(libopenage
 	object.cpp
 	render_entity.cpp
 	render_stage.cpp
-	shader_template.cpp
 )
